# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant SaaS POS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory levels
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: productId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustment"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List stock movements
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: productId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"

  # Customers
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard KPI summary
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: period, in: query, schema: { type: string, enum: [today, week, month, year] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/sales-chart:
    get:
      operationId: getSalesChart
      tags: [dashboard]
      summary: Get sales trend data for chart
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: period, in: query, schema: { type: string, enum: [week, month, year] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesDataPoint"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Get recent orders
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/low-stock-alerts:
    get:
      operationId: getLowStockAlerts
      tags: [dashboard]
      summary: Get low stock products
      parameters:
        - { name: branchId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowStockAlert"

  # Reports
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [id, productId, productName, branchId, quantity, updatedAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productSku: { type: ["string", "null"] }
        branchId: { type: integer }
        branchName: { type: string }
        quantity: { type: integer }
        reorderPoint: { type: ["integer", "null"] }
        updatedAt: { type: string, format: date-time }
    InventoryAdjustment:
      type: object
      required: [productId, branchId, quantity, reason]
      properties:
        productId: { type: integer }
        branchId: { type: integer }
        quantity: { type: integer }
        reason: { type: string, enum: [purchase, sale, adjustment, return, damage, transfer] }
        notes: { type: string }
    InventoryMovement:
      type: object
      required: [id, productId, productName, branchId, quantity, reason, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        branchId: { type: integer }
        branchName: { type: string }
        quantity: { type: integer }
        reason: { type: string }
        notes: { type: ["string", "null"] }
        referenceId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }

    # Customer
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        discount: { type: number }
        totalPrice: { type: number }
        notes: { type: ["string", "null"] }
    Order:
      type: object
      required: [id, orderNumber, branchId, status, subtotal, total, createdAt]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        branchId: { type: integer }
        branchName: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        cashierId: { type: integer }
        cashierName: { type: string }
        status: { type: string, enum: [draft, held, completed, refunded, voided] }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: ["string", "null"] }
        amountPaid: { type: ["number", "null"] }
        change: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"] }
...
    DashboardSummary:
      type: object
      required: [totalRevenue, totalOrders, totalCustomers, totalProducts, avgOrderValue, revenueChange, ordersChange]
      properties:
        totalRevenue: { type: number }
        totalOrders: { type: integer }
        totalCustomers: { type: integer }
        totalProducts: { type: integer }
        avgOrderValue: { type: number }
        revenueChange: { type: number }
        ordersChange: { type: number }
        customersChange: { type: number }
        lowStockCount: { type: integer }
        pendingRefunds: { type: integer }
    SalesDataPoint:
      type: object
      required: [label, revenue, orders]
      properties:
        label: { type: string }
        revenue: { type: number }
        orders: { type: integer }
    TopProduct:
      type: object
      required: [productId, productName, totalSold, revenue]
      properties:
        productId: { type: integer }
        productName: { type: string }
        categoryName: { type: ["string", "null"] }
        totalSold: { type: integer }
        revenue: { type: number }
    LowStockAlert:
      type: object
      required: [productId, productName, branchId, branchName, currentStock, reorderPoint]
      properties:
        productId: { type: integer }
        productName: { type: string }
        branchId: { type: integer }
        branchName: { type: string }
        currentStock: { type: integer }
        reorderPoint: { type: integer }

    # Reports
...